pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurosynth"
version = "2.0.0"
description = "Unified Neurosurgical Knowledge Extraction & Retrieval System (Phase 1 + Phase 2)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "NeuroSynth Team"}
]
keywords = ["neurosurgery", "medical", "nlp", "rag", "knowledge-extraction", "retrieval"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

# Core dependencies (Phase 1 + Phase 2)
dependencies = [
    # PDF Processing (Phase 1)
    "PyMuPDF>=1.23.0",

    # Core utilities
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",

    # Database (Phase 2)
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",

    # API (Phase 2)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",

    # Configuration (Phase 2)
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
# Vector search (Phase 2)
faiss = ["faiss-cpu>=1.7.4"]
faiss-gpu = ["faiss-gpu>=1.7.4"]

# Embeddings providers (Phase 1 + Phase 2)
voyage = ["voyageai>=0.2.0"]
openai = ["openai>=1.0.0"]
local = ["sentence-transformers>=2.2.0"]
biomedclip = ["torch>=2.0.0", "open-clip-torch>=2.20.0"]
clip = ["torch>=2.0.0", "clip @ git+https://github.com/openai/CLIP.git"]

# LLM (Phase 2 RAG)
anthropic = ["anthropic>=0.40.0"]

# Phase 1 enhancements
graph = ["networkx>=3.0"]
ocr = ["pytesseract>=0.3.10"]
vlm = ["anthropic>=0.7.0"]
colbert = ["colbert-ai>=0.2.0", "torch>=2.0.0"]
config = ["pyyaml>=6.0"]

# Re-ranking (Phase 2)
reranking = ["sentence-transformers>=2.2.0"]

# Development
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# All features
all = [
    "faiss-cpu>=1.7.4",
    "voyageai>=0.2.0",
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "open-clip-torch>=2.20.0",
    "anthropic>=0.40.0",
    "networkx>=3.0",
    "pytesseract>=0.3.10",
    "pyyaml>=6.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/neurosynth/neurosynth-v2"
Documentation = "https://github.com/neurosynth/neurosynth-v2#readme"
Repository = "https://github.com/neurosynth/neurosynth-v2"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=html"
